+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - drones can now use status codes to express ceratin things
 - added ai_status as a command to get debug-information about the AI
+- added hive-coordination and elite-direction as channels
 
 ### Changed
 - hive-playroom is now hive-play-room